Reorganized validation module and functions used in the validation
